+++ src/gwt/src/org/rstudio/studio/client/workbench/views/source/editors/text/TextEditingTarget.java
@@ -2052,11 +2052,6 @@ public class TextEditingTarget implements
    @Handler
    void onInsertRoxygenSkeleton()
    {
-      if (!DocumentMode.isSelectionInRMode(docDisplay_))
-      {
-         showRModeWarning("Roxygen Skeleton");
-         return;
-      }
       docDisplay_.insertRoxygenSkeleton();
    }